:recycle: (filters) extract connector, adapter and api in their folder
