+++ frontend/src/modules/filters/connector/connector.ts
@@ -1,6 +1,6 @@
 import { adaptDifficulties } from './adapter';
 import { fetchDifficulties } from './api';
-import { Choices } from './interface';
+import { Choices } from '../interface';
 
 export const getDifficulties = async (): Promise<Choices> => {
   const rawDifficulties = await fetchDifficulties({ language: 'fr' });
