+++ frontend/src/modules/filters/connector/index.ts
@@ -1,0 +1,1 @@
+export { getDifficulties } from './connector';
